backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlmodel==0.0.18
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1
